requirements.txt
bottle==0.12.23
CherryPy==8.9.0
pyserial==3.5
PyYAML
